// src/store/user/user.types.ts
 export enum USER_ACTION_TYPES {
    SET_CURRENT_USER = 'SET_CURRENT_USER',
    SET_CURRENT_USER_START = 'SET_CURRENT_USER_START',
    CHECK_USER_SESSION  = 'CHECK_USER_SESSION',
    GOOGLE_SIGN_IN_START  = 'GOOGLE_SIGN_IN_START',
    EMAIL_SIGN_IN_START  = 'EMAIL_SIGN_IN_START',
    SIGN_IN_SUCCESS  = 'SIGN_IN_SUCCESS',
    SIGN_IN_FAILED  = 'SIGN_IN_FAILED',
    SET_LOG_OUT_START = 'SET_LOG_OUT_START',
    SET_LOG_OUT_SUCCESS = 'SET_LOG_OUT_SUCCESS',
    SET_LOG_OUT_FAILED = 'SET_LOG_OUT_FAILED',
    SIGN_UP_START = 'SIGN_UP_START',
    SIGN_UP_SUCCESS  = 'SIGN_UP_SUCCESS',
    SIGN_UP_FAILED  = 'SIGN_UP_FAILED'
}

export type User = {
    email: string;
    displayName: string;
    password: string;
    createdAt: string;
    id: number;
}